Refactor fitness evaluation and scheduling logic
- Improved array operations in `Schedule` class for better readability and performance.
- Added new methods in `Schedule` to retrieve round counts needed.
- Enhanced `StableConfigHash` to use canonical representations for rounds and requirements.
- Updated `FitnessBase` to streamline initialization and configuration handling.
- Refactored `FitnessEvaluator` classes to utilize numpy's built-in functions for efficiency.
- Introduced a new CLI for the genetic algorithm scheduler, including commands for listing, setting, and adding configurations.
- Implemented logging configuration for better debugging and monitoring.
- Added hard constraint checking classes to validate scheduling constraints.
- Improved crossover and selection operations in genetic algorithms for clarity and performance.

+++ fll_scheduler_ga/operators/selection.py
@@ -55,7 +55,7 @@ class RandomSelect(Selection):
             while i1 == i2:
                 i2 = self.rng.integers(0, n)
 
-            return np.array([i1, i2], dtype=np.int16)
+            return np.array([i1, i2], dtype=int)
 
         choices = np.arange(n)
         self.rng.shuffle(choices)
